+++ src/main/kotlin/com/aaronhowser1/ariadnesthread/AriadnesThread.kt
@@ -43,6 +43,7 @@ object AriadnesThread {
     }
 
     private fun onClientSetup(event: FMLClientSetupEvent) {
+        ClientConfig.checkValid()
     }
 
     private fun onServerSetup(event: FMLDedicatedServerSetupEvent) {